requirements.txt
PyYaml
starlette
openapi_core
requests
uvicorn